third_party/WebKit/LayoutTests/fast/dom/DOMImplementation/createDocumentType-null.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var doctype = document.implementation.createDocumentType(null, null, null);
shouldBeEqualToString("doctype.name", "null");
shouldBeEqualToString("doctype.publicId", "null");
shouldBeEqualToString("doctype.systemId", "null");
</script>
</body>
</html>
