third_party/blink/web_tests/platform/generic/fast/dom/DOMImplementation/createDocumentType-null-expected.txt

PASS doctype.name is "null"
PASS doctype.publicId is "null"
PASS doctype.systemId is "null"
PASS successfullyParsed is true

TEST COMPLETE
